The Final Project File Upload of CovidNinja v1.0
Thus, this is our final project file of Covid-Ninja v1.0.
This is Game Filter, where the user will be able to control the Game Via his/her Eye Blinks and the Ninja will jump over the Covid-19 model and also will show precautionary measures against corona pandemic and has the branding of the company 'HealthRx'.